ensembl ID to Symbol
Ensembl and GenAnT use gene IDs with gene symbols. As a default we were transferring gene IDs over, making finding the gene symbol in the gene_symbol_table inconvenient. We added an R script to processing the reference directory so that gene symbols get transferred instead of GeneIDs when possible. It also deals with redundant gene symbols and pseudogenes.